etl/requirements.txt
redis>=4.5.0
asyncpg==0.27.0
python-dateutil==2.8.2
python-dotenv==1.0.0
